vuexの導入ができない

+++ laravel/laravel/database/migrations/2022_04_25_154357_add_review_key_to_bookings_table.php
@@ -25,8 +25,10 @@ class AddReviewKeyToBookingsTable extends Migration
      */
     public function down()
     {
-        Schema::table('bookings', function (Blueprint $table) {
-            $table->dropColumn('review_key');
-        });
+        if (Schema::hasColumn('bookings', 'review_key')) {
+              Schema::table('bookings', function (Blueprint $table) {
+                $table->dropColumn('review_key');
+              });
+            }
     }
 }
